requirements.txt
# Core Dependencies
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# AI APIs
anthropic>=0.18.0
elevenlabs>=0.2.0
requests>=2.31.0

# Video Processing
moviepy>=1.0.3
opencv-python>=4.8.0
pillow>=10.0.0

# Audio Processing
pydub>=0.25.1

# CLI & Logging
rich>=13.0.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
